Disastrous surrender of day 14

+++ src/main/scala/uk/co/hadoopathome/adventofcode19/day14/package.scala
@@ -1,5 +1,5 @@
 package uk.co.hadoopathome.adventofcode19
 
 package object day14 {
-case class Equation(lhs: Map[String, Long], rhs: (String, Long))
+case class Equation(lhs: Map[String, Double], rhs: (String, Double))
 }
